+++ lib/toxiproxy.rb
@@ -1,7 +1,10 @@
 require "json"
 require "uri"
 require "net/http"
-require "forwardable"
+
+require "toxiproxy/collection"
+require "toxiproxy/toxic"
+require "toxiproxy/toxic_collection"
 
 class Toxiproxy
   URI = ::URI.parse("http://127.0.0.1:8474")
@@ -10,107 +13,6 @@ class Toxiproxy
   class NotFound < StandardError; end
   class ProxyExists < StandardError; end
   class InvalidToxic < StandardError; end
-
-  # ProxyCollection represents a set of proxies. This allows to easily perform
-  # actions on every proxy in the collection.
-  #
-  # Unfortunately, it doesn't implement all of Enumerable because there's no way
-  # to subclass an Array or include Enumerable for the methods to return a
-  # Collection instead of an Array (see MRI). Instead, we delegate methods where
-  # it doesn't matter and only allow the filtering methods that really make
-  # sense on a proxy collection.
-  class Collection
-    extend Forwardable
-
-    DELEGATED_METHODS = [:length, :size, :count, :find, :each, :map]
-    DEFINED_METHODS   = [:select, :reject, :grep, :down]
-    METHODS = DEFINED_METHODS + DELEGATED_METHODS
-
-    def_delegators :@collection, *DELEGATED_METHODS
-
-    def initialize(collection)
-      @collection = collection
-    end
-
-    # Sets every proxy in the collection as down. For example:
-    #
-    #   Toxiproxy.grep(/redis/).down { .. }
-    #
-    # Would simulate every Redis server being down for the duration of the
-    # block.
-    def down(*args, &block)
-      @collection.inject(block) { |nested, proxy|
-        -> { proxy.down(*args, &nested) }
-      }.call
-    end
-
-    # Destroys all toxiproxy's in the collection
-    def destroy
-      @collection.each(&:destroy)
-    end
-
-    def select(&block)
-      self.class.new(@collection.select(&block))
-    end
-
-    def reject(&block)
-      self.class.new(@collection.reject(&block))
-    end
-
-    # Grep allows easily selecting a subset of proxies, by returning a
-    # ProxyCollection with every proxy name matching the regex passed.
-    def grep(regex)
-      self.class.new(@collection.select { |proxy|
-        proxy.name =~ regex
-      })
-    end
-  end
-
-  class Toxic
-    attr_reader :name, :proxy, :direction
-    attr_reader :attrs
-
-    def initialize(proxy:, name:, direction:, attrs: {})
-      @proxy     = proxy
-      @name      = name
-      @direction = direction
-      @attrs     = attrs
-    end
-
-    def enabled?
-      attrs[:enabled]
-    end
-
-    def enable
-      attrs[:enabled] = true
-      save
-    end
-
-    def disable
-      attrs[:enabled] = false
-      save
-    end
-
-    def []=(name, value)
-      attrs[name] = value
-    end
-
-    def save
-      unless VALID_DIRECTIONS.include?(direction.to_sym)
-        raise InvalidToxic, "Toxic direction must be one of: [#{VALID_DIRECTIONS.join(', ')}], got: #{direction}"
-      end
-      request = Net::HTTP::Post.new("/proxies/#{proxy.name}/#{direction}/toxics/#{name}")
-
-      request.body = attrs.to_json
-
-      response = Toxiproxy.http.request(request)
-      assert_response(response)
-
-      @attrs = JSON.parse(response.body)
-
-      self
-    end
-  end
 
   attr_reader :listen, :name
 
@@ -168,47 +70,7 @@ class Toxiproxy
     find_by_name(query)
   end
 
-  class ToxicCollection
-    extend Forwardable
-
-    attr_accessor :toxics
-    attr_reader :proxy
-
-    def_delegators :@toxics, :<<, :find
-
-    def initialize(proxy)
-      @proxy  = proxy
-      @toxics = []
-    end
-
-    def apply(&block)
-      @toxics.each(&:enable)
-      yield
-    ensure
-      @toxics.each(&:disable)
-    end
-
-    def upstream(toxic_name, attrs = {})
-      toxics << Toxic.new(
-        name: toxic_name,
-        proxy: proxy,
-        direction: :upstream,
-        attrs: attrs
-      )
-      self
-    end
-
-    def downstream(toxic_name, attrs = {})
-      toxics << Toxic.new(
-        name: toxic_name,
-        proxy: proxy,
-        direction: :downstream,
-        attrs: attrs
-      )
-      self
-    end
-  end
-
+  # Set an upstream toxic.
   def upstream(toxic = nil, attrs = {})
     return @upstream unless toxic
 
@@ -217,6 +79,7 @@ class Toxiproxy
     collection
   end
 
+  # Set a downstream toxic.
   def downstream(toxic, attrs = {})
     collection = ToxicCollection.new(self)
     collection.downstream(toxic, attrs)
@@ -296,17 +159,21 @@ class Toxiproxy
   def http
     self.class.http
   end
-end
 
-def assert_response(response)
-  case response
-  when Net::HTTPConflict
-    raise Toxiproxy::ProxyExists, response.body
-  when Net::HTTPNotFound
-    raise Toxiproxy::NotFound, response.body
-  when Net::HTTPBadRequest
-    raise Toxiproxy::InvalidToxic, response.body
-  else
-    response.value # raises if not OK
+  def self.assert_response(response)
+    case response
+    when Net::HTTPConflict
+      raise Toxiproxy::ProxyExists, response.body
+    when Net::HTTPNotFound
+      raise Toxiproxy::NotFound, response.body
+    when Net::HTTPBadRequest
+      raise Toxiproxy::InvalidToxic, response.body
+    else
+      response.value # raises if not OK
+    end
+  end
+
+  def assert_response(*args)
+    self.class.assert_response(*args)
   end
 end
